renamed unclear check method
